refactor nostr hook files
