+++ components/PubkeyProfile.tsx
@@ -2,8 +2,8 @@ import { brandColors } from "@/constants";
 import { View, Image, TouchableOpacity } from "react-native";
 import { useState } from "react";
 import { openURL } from "expo-linking";
-import { NostrUserProfile, useAddFollower, useRemoveFollower } from "@/utils";
-import { useAuth, useUser } from "@/hooks";
+import { NostrUserProfile } from "@/utils";
+import { useAddFollower, useAuth, useRemoveFollower, useUser } from "@/hooks";
 import Icon from "react-native-vector-icons/MaterialIcons";
 import { useTheme } from "@react-navigation/native";
 import { useRouter } from "expo-router";
