+++ hooks/nostrProfile/index.ts
@@ -1,2 +1,4 @@
 export * from "./useNostrProfile";
 export * from "./useSaveNostrProfile";
+export * from "./useAddFollower";
+export * from "./useRemoveFollower";
